+++ components/TextFields/src/ColorThemer/MDCFilledTextFieldColorThemer.m
@@ -14,12 +14,12 @@
 
 #import "MDCFilledTextFieldColorThemer.h"
 
-static CGFloat const kFilledTextFieldActiveAlpha = 0.87f;
-static CGFloat const kFilledTextFieldOnSurfaceAlpha = 0.6f;
-static CGFloat const kFilledTextFieldDisabledAlpha = 0.38f;
-static CGFloat const kFilledTextFieldSurfaceOverlayAlpha = 0.04f;
-static CGFloat const kFilledTextFieldIndicatorLineAlpha = 0.42f;
-static CGFloat const kFilledTextFieldIconAlpha = 0.54f;
+static CGFloat const kFilledTextFieldActiveAlpha = (CGFloat)0.87;
+static CGFloat const kFilledTextFieldOnSurfaceAlpha = (CGFloat)0.6;
+static CGFloat const kFilledTextFieldDisabledAlpha = (CGFloat)0.38;
+static CGFloat const kFilledTextFieldSurfaceOverlayAlpha = (CGFloat)0.04;
+static CGFloat const kFilledTextFieldIndicatorLineAlpha = (CGFloat)0.42;
+static CGFloat const kFilledTextFieldIconAlpha = (CGFloat)0.54;
 
 @implementation MDCFilledTextFieldColorThemer
