+++ components/private/ThumbTrack/src/MDCThumbTrack.m
@@ -24,18 +24,18 @@
 
 #pragma mark - ThumbTrack constants
 
-static const CGFloat kAnimationDuration = 0.25f;
-static const CGFloat kThumbChangeAnimationDuration = 0.12f;
-static const CGFloat kDefaultThumbBorderWidth = 2.0f;
-static const CGFloat kDefaultThumbRadius = 6.0f;
-static const CGFloat kDefaultTrackHeight = 2.0f;
+static const CGFloat kAnimationDuration = (CGFloat)0.25;
+static const CGFloat kThumbChangeAnimationDuration = (CGFloat)0.12;
+static const CGFloat kDefaultThumbBorderWidth = 2;
+static const CGFloat kDefaultThumbRadius = 6;
+static const CGFloat kDefaultTrackHeight = 2;
 static const CGFloat kDefaultFilledTrackAnchorValue = -CGFLOAT_MAX;
-static const CGFloat kTrackOnAlpha = 0.5f;
-static const CGFloat kMinTouchSize = 48.0f;
-static const CGFloat kThumbSlopFactor = 3.5f;
-static const CGFloat kValueLabelHeight = 48.f;
-static const CGFloat kValueLabelWidth = 0.81f * kValueLabelHeight;
-static const CGFloat kValueLabelFontSize = 12.f;
+static const CGFloat kTrackOnAlpha = (CGFloat)0.5;
+static const CGFloat kMinTouchSize = 48;
+static const CGFloat kThumbSlopFactor = (CGFloat)3.5;
+static const CGFloat kValueLabelHeight = 48;
+static const CGFloat kValueLabelWidth = (CGFloat)0.81 * kValueLabelHeight;
+static const CGFloat kValueLabelFontSize = 12;
 
 static UIColor *ValueLabelTextColorDefault() {
   return UIColor.whiteColor;
@@ -109,8 +109,8 @@ static UIColor *InkColorDefault() {
 
     // Allow an extra 10% of the increment to guard against rounding errors excluding dots that
     // should genuinely be within the active segment.
-    CGFloat minActiveX = CGRectGetMinX(self.activeDotsSegment) - relativeIncrement * 0.1f;
-    CGFloat maxActiveX = CGRectGetMaxX(self.activeDotsSegment) + relativeIncrement * 0.1f;
+    CGFloat minActiveX = CGRectGetMinX(self.activeDotsSegment) - relativeIncrement * (CGFloat)0.1;
+    CGFloat maxActiveX = CGRectGetMaxX(self.activeDotsSegment) + relativeIncrement * (CGFloat)0.1;
     for (NSUInteger i = 0; i < _numDiscreteDots; i++) {
       CGFloat relativePosition = i * relativeIncrement;
       if (minActiveX <= relativePosition && maxActiveX >= relativePosition) {
@@ -377,7 +377,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
     _valueLabel = [[MDCNumericValueLabel alloc]
         initWithFrame:CGRectMake(0, 0, kValueLabelWidth, kValueLabelHeight)];
     // Effectively 0, but setting it to 0 results in animation not happening
-    _valueLabel.transform = CGAffineTransformMakeScale(0.001f, 0.001f);
+    _valueLabel.transform = CGAffineTransformMakeScale((CGFloat)0.001, (CGFloat)0.001);
     _valueLabel.fontSize = kValueLabelFontSize;
     [self addSubview:_valueLabel];
   } else {
@@ -624,7 +624,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
       void (^afterCrossingAnchorAnimation)(BOOL) = ^void(__unused BOOL finished) {
         UIViewAnimationOptions options = baseAnimationOptions | UIViewAnimationOptionCurveEaseOut;
         [UIView animateWithDuration:(kAnimationDuration - animationDurationToAnchor)
-                              delay:0.0f
+                              delay:0
                             options:options
                          animations:^{
                            [self updateViewsMainIsAnimated:animated
@@ -636,7 +636,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
       };
       UIViewAnimationOptions options = baseAnimationOptions | UIViewAnimationOptionCurveEaseIn;
       [UIView animateWithDuration:animationDurationToAnchor
-                            delay:0.0f
+                            delay:0
                           options:options
                        animations:^{
                          self.value = self.filledTrackAnchorValue;
@@ -648,7 +648,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
                        completion:afterCrossingAnchorAnimation];
     } else {
       [UIView animateWithDuration:kAnimationDuration
-                            delay:0.0f
+                            delay:0
                           options:baseAnimationOptions
                        animations:^{
                          if (activeSegmentShrinking) {
@@ -661,9 +661,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
                        completion:animationCompletion];
     }
   } else {
-    [self updateViewsMainIsAnimated:animated
-                       withDuration:0.0f
-                   animationOptions:baseAnimationOptions];
+    [self updateViewsMainIsAnimated:animated withDuration:0 animationOptions:baseAnimationOptions];
     [self updateDotsViewActiveSegment];
     [self updateThumbAfterMoveAnimated:animateThumbAfterMove
                                options:baseAnimationOptions
@@ -676,7 +674,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
                           completion:(void (^)(void))completion {
   if (animated) {
     [UIView animateWithDuration:kThumbChangeAnimationDuration
-        delay:0.0f
+        delay:0
         options:animationOptions
         animations:^{
           [self updateViewsForThumbAfterMoveIsAnimated:animated
@@ -688,7 +686,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
           }
         }];
   } else {
-    [self updateViewsForThumbAfterMoveIsAnimated:animated withDuration:0.0f];
+    [self updateViewsForThumbAfterMoveIsAnimated:animated withDuration:0];
 
     if (completion) {
       completion();
@@ -830,7 +828,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
     if (self.enabled && _isDraggingThumb) {
       _valueLabel.transform = CGAffineTransformIdentity;
     } else {
-      _valueLabel.transform = CGAffineTransformMakeScale(0.001f, 0.001f);
+      _valueLabel.transform = CGAffineTransformMakeScale((CGFloat)0.001, (CGFloat)0.001);
     }
   }
 
@@ -1018,7 +1016,7 @@ static inline CGFloat DistanceFromPointToPoint(CGPoint point1, CGPoint point2) {
 
   CGFloat scaledTargetValue = (targetValue - _minimumValue) / (_maximumValue - _minimumValue);
   CGFloat snappedValue =
-      MDCRound((_numDiscreteValues - 1) * scaledTargetValue) / (_numDiscreteValues - 1.0f);
+      MDCRound((_numDiscreteValues - 1) * scaledTargetValue) / (_numDiscreteValues - 1);
   return (1 - snappedValue) * _minimumValue + snappedValue * _maximumValue;
 }
